fix strlcat strlcpy

+++ libft/ft_strlcat.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ttangcha <ttangcha@student.42.fr>          +#+  +:+       +#+        */
+/*   By: thanaphat <thanaphat@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/04 18:26:34 by ttangcha          #+#    #+#             */
-/*   Updated: 2024/09/07 19:09:13 by ttangcha         ###   ########.fr       */
+/*   Updated: 2024/09/17 19:02:45 by thanaphat        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	unsigned int	dstsize;
-	unsigned int	srcsize;
-	unsigned int	i;
-	unsigned int	j;
+	size_t	dstsize;
+	size_t	srcsize;
+	size_t	i;
+	size_t	j;
 
 	j = ft_strlen(dst);
 	dstsize = ft_strlen(dst);
